# openapi.yaml
openapi: 3.1.0
info:
  title: Extracto v1 API
  version: 1.0.0
  description: |
    Stable bearer-auth HTTP surface for Extracto. Mint a key with
    `extracto api-key create <user-email> <name>` (or `POST /api/v1/keys`
    via session auth), then send it as `Authorization: Bearer <key>` on
    every call below.

    Mutation endpoints additionally require a same-origin `Origin` header
    that matches the request URL (CSRF-style).
  license:
    name: MIT
    url: https://github.com/codelined-ag/extracto/blob/main/LICENSE
servers:
  - url: http://localhost:3000
    description: Local Docker
  - url: https://extracto.example.com
    description: Production
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: extr_*
    metricsBearer:
      type: http
      scheme: bearer
      description: Separate token, set via `METRICS_TOKEN` env.
  schemas:
    CloudWatcher:
      type: object
      required: [id, userId, provider, name, intervalSeconds, active, model, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        provider: { type: string, enum: [dropbox, google_drive, onedrive, local] }
        name: { type: string }
        folderPath: { type: string }
        intervalSeconds: { type: integer }
        active: { type: boolean }
        model: { type: string }
        templateId: { type: string, nullable: true }
        autoKbExport: { type: boolean }
        autoS3Export: { type: boolean }
        lastPolledAt: { type: string, format: date-time, nullable: true }
        lastError: { type: string, nullable: true }
        consecutiveFailures: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
security:
  - bearer: []
paths:
...
  /api/v1/integrations/watchers:
    get:
      summary: List the user's cloud watched folders
      operationId: listCloudWatchers
      tags: [Integrations]
      responses:
        "200":
          description: List of watched folders
          content:
            application/json:
              schema:
                type: object
                properties:
                  watchers:
                    type: array
                    items: { $ref: "#/components/schemas/CloudWatcher" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create a cloud watched folder
      operationId: createCloudWatcher
      tags: [Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, name, model]
              properties:
                provider: { type: string, enum: [dropbox, google_drive, onedrive, local] }
                name: { type: string, maxLength: 80 }
                folderPath: { type: string, default: "" }
                model: { type: string }
                intervalSeconds:
                  type: integer
                  minimum: 60
                  maximum: 86400
                  default: 300
                active: { type: boolean, default: true }
                templateId: { type: string, nullable: true }
                autoKbExport: { type: boolean, default: false }
                autoS3Export: { type: boolean, default: false }
      responses:
        "201":
          description: Watcher created
          content:
            application/json:
              schema:
                type: object
                properties:
                  watcher: { $ref: "#/components/schemas/CloudWatcher" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: A watcher with that name already exists for this provider
  /api/v1/integrations/watchers/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    patch:
      summary: Update a cloud watched folder
      operationId: updateCloudWatcher
      tags: [Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                folderPath: { type: string }
                model: { type: string }
                intervalSeconds: { type: integer, minimum: 60, maximum: 86400 }
                active: { type: boolean }
                templateId: { type: string, nullable: true }
                autoKbExport: { type: boolean }
                autoS3Export: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  watcher: { $ref: "#/components/schemas/CloudWatcher" }
        "404":
          description: Watcher not found
    delete:
      summary: Delete a cloud watched folder
      operationId: deleteCloudWatcher
      tags: [Integrations]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  removed: { type: boolean }
        "404":
          description: Watcher not found
  /api/v1/integrations/{provider}/oauth-app:
    parameters:
      - { name: provider, in: path, required: true, schema: { type: string, enum: [dropbox, google_drive, onedrive] } }
    get:
      summary: Read the user's saved OAuth app credentials status
      operationId: getOauthApp
      tags: [Integrations]
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider: { type: string }
                  source: { type: string, enum: [user, server, none] }
                  clientIdLast4: { type: string, nullable: true }
                  redirectUri: { type: string }
    put:
      summary: Save the user's own OAuth client_id + client_secret
      operationId: putOauthApp
      tags: [Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId, clientSecret]
              properties:
                clientId: { type: string, maxLength: 200 }
                clientSecret: { type: string, maxLength: 400 }
      responses:
        "200":
          description: Saved
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      summary: Clear the user's saved OAuth credentials (falls back to server-wide)
      operationId: deleteOauthApp
      tags: [Integrations]
      responses:
        "200":
          description: Deleted